requirements.txt
dill
fastapi
numpy
pydantic_settings
python-multipart
uvicorn